Moves library and library tests into src/lib/
