+++ src/lib/__tests__/componentConstruction.test.js
