+++ src/lib/motionController.js
